### 2025-11-18 #9461 @editorialbot
[PRE REVIEW]: datadriftR: an R package for streaming data drift detection
**Submitting author:** @ugurdar (<a href="http://orcid.org/0009-0005-8076-2199">Ugur Dar</a>)
**Repository:** https://github.com/ugurdar/datadriftR
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.1.0
