### 2025-11-20 #9481 @editorialbot
[REVIEW]: datadriftR: an R package for streaming data drift detection
**Submitting author:** @ugurdar (<a href="http://orcid.org/0009-0005-8076-2199">Ugur Dar</a>)
**Repository:** https://github.com/ugurdar/datadriftR
**Branch with paper.md** (empty if default branch): paper
**Version:** 1.1.0
