### 2025-11-21 #9481 @benyamindsmith
👋 @crvernon

I couldn't help myself but start the review process by opening several issues related to documentation and the code. Please see my issues here: https://github.com/ugurdar/datadriftR/issues
