### 2025-11-21 ugurdar/datadriftR#9 @fangzhou-xie
[JOSS reivew]: report
I have finished the review of `datadriftR` package [submission at JOSS](https://github.com/openjournals/joss-reviews/issues/9481), and here is my report.

# 1. Summary

This package implements multiple algorithms in online learning for drift detection in R with R6 classes.

# 2. Potential Issues

Here are some issues that I identified,  sorted by the order of the review checklist.

## 1. Substantial scholarly effort

Most methods implemented in this paper seem to be the literal translation of the [drift detection module of `scikit-multiflow`](https://scikit-multiflow.readthedocs.io/en/stable/index.html), except for the `ProfileDifference` class, which I believe is implemented from the ground up by the authors. The `ProfileDifference` class is less than 200 lines, with the total LOC for this package is a little shy of 1K.

I am not entirely sure about the "substantial scholarly effort" criterion, and ultimately it's a decision by the editor @crvernon.

## 2. Installation issues

As also pointed out by the other reviewer (#2, #3, #7):

1. the installation instruction needs to be updated (from `"yourusername/datadriftR"` to `"ugurdar/datadriftR"`)
2. the README should reflect the CRAN installation, given that the package is already published there (e.g. `install.packages("datadriftR")`)
3. dependencies not listed in "Imports" in the DESCRIPTION file. This is the reason why the other reviewer and I couldn't install the package from Github. I noticed that, however,  the [CRAN version of the package](https://github.com/cran/datadriftR) did have `fda.usc` and `doremi` listed under "Imports". So it seems to me that the authors published the package from a local version to CRAN, but didn't commit and push the development version on Github.
4. version control issue. Following the previous note, the development version on Github is version 0.1.0, but the CRAN one is 1.0.0. Usually speaking the development version should always be at least as high as the published version, but never lower. This confirms that the authors did modifications to the local version while preparing for the CRAN release, but didn't update the Github repo.

## 3. Functionality

Related to the discussion #5 of package functionality in R, I can understand that the choice made at `scikit-multiflow` for a `for-loop` based detection, as they are aiming for online learning even possibly in real time, e.g. subscribing to a web stream and processing on the fly. But I think this use case is not as prevalent for R as for Python, as the two communities are fundamentally different. I hence agree with the other reviewer that it is better to package the drift detection methods into functions that take vectors/lists/dataframes and report the detection result as output.

But I am not saying to get rid of the "online learning" abilities: if the authors still want to keep the `for-loop` online learning functionality, those methods could be "another way to use the package".

## 4. Documentation

1. installation documentation issue (mentioned in the installation issue above)
2. dependency issue (also above)
3. online documentation is mentioned in the paper but can't be found  (see also #8). Using `pkgdown` to create a website for online documentation is very simple, especially I notice that the package is already well documented with the Roxygen for the functions.
4. example missing for KSWIN #6
5. no automated test found. Especially most functions are directly translated from `scikit-multiflow`, adding tests to confirm the implementations yield identical results to the Python package is important. For example, use `tinytest` or `testthat` for that purpose.

## 5. Misc

1. lightweight: The README file listed "lightweight" as one of the features for this package, "depend only on base R and R6". First of all, `fda.usc` and `doremi` are used but not declared in DESCRIPTION; second, both packages have a fair number of dependencies themselves. Therefore, I think the claim of "lightweight" is somewhat misleading as a feature.
2. related the functionality comment on packaging the methods into functions: it would be easier for users if the output results are in some format that is manipulatable, rather than printing out messages in the console. For example, having an output of dataframe with each row being the  detected drift, with index and other useful statistics etc.
