### 2025-12-05 ugurdar/datadriftR#9 @fangzhou-xie
@ugurdar Thanks for your work on this package.

It is to my understanding that you might need to address the concerns/comments that the reviewers made by:
1. try to modify/update the code/documentation for your package;
2. or in the case you believe that any specific comment is not true or we misunderstood anything in the paper or the code, you can write a "rebuttal" to explain/argue that the reviewer's comment is not true/misunderstood.

Unless you disagree with any of the comments, you could start updating your package (in your package repo) by directly addressing our comments. Otherwise, feel free to further comment here if you have any question on the points I listed (I believe the same applies to the other reviewer).

After everything is fixed, we (as reviewers) would then go to the [JOSS review issue](https://github.com/openjournals/joss-reviews/issues/9481) and finish our report checklist. Then the editor would decide whether or not to accept your JOSS paper.
