+++ vignettes/datadriftR-intro.Rmd
@@ -12,179 +12,164 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
   fig.width = 7,
-  fig.height = 4
+  fig.height = 4,
+  message = FALSE,
+  warning = FALSE
 )
 ```
 
-## What is Concept Drift?
+## What is Data Drift?
 
-In machine learning systems, **concept drift** occurs when the statistical properties of the data change over time. This can cause model performance to degrade because the patterns learned during training no longer apply to new data.
-
-For example:
-- A spam filter trained on old emails may fail to catch new spam techniques
-- A fraud detection model may miss new types of fraudulent behavior
-- A recommendation system may become less accurate as user preferences evolve
-
-**datadriftR** helps you detect these changes in real-time by monitoring your data streams and alerting you when drift occurs.
-
-## Quick Start
-
-The `detect_drift()` function is the easiest way to detect drift in your data:
-
-```{r quickstart}
-library(datadriftR)
-set.seed(123)
-
-# Create a data stream with a sudden change at position 501
-# First 500 observations: 30% ones, 70% zeros
-# Last 500 observations: 70% ones, 30% zeros (drift!)
-stable_period <- sample(c(0, 1), 500, replace = TRUE, prob = c(0.7, 0.3))
-drift_period  <- sample(c(0, 1), 500, replace = TRUE, prob = c(0.3, 0.7))
-stream <- c(stable_period, drift_period)
-
-# Detect drift using DDM method
-results <- detect_drift(stream, method = "ddm")
-print(results)
-```
-
-The function returns a data frame with all detected drift points and their positions in the stream.
+Data drift occurs when the statistical properties of data change over time. This is critical in machine learning systems where models may degrade as data patterns shift.
 
 ## Available Methods
 
-datadriftR provides 8 different drift detection methods. Each method has its own strengths:
+**datadriftR** provides 8 drift detection methods:
 
-| Method | Best For | Description |
-|--------|----------|-------------|
-| `ddm` | Binary streams | Monitors error rate changes |
-| `eddm` | Binary streams | Detects drift earlier than DDM |
-| `hddm_a` | Binary streams | Uses Hoeffding bounds (averaging) |
-| `hddm_w` | Binary streams | Uses Hoeffding bounds (weighted) |
-| `kswin` | Any numeric data | Kolmogorov-Smirnov statistical test |
-| `page_hinkley` | Any numeric data | Cumulative sum approach |
-| `kl_divergence` | Any numeric data | Information-theoretic approach |
-| `profile_difference` | Functional data | Compares data profiles |
+| Method | Best For | Key Advantage |
+|:-------|:---------|:--------------|
+| [DDM](../reference/DDM.html) | Binary error streams | Simple and fast |
+| [EDDM](../reference/EDDM.html) | Binary error streams | Earlier detection than DDM |
+| [HDDM-A](../reference/HDDM_A.html) | Binary streams | Theoretical guarantees |
+| [HDDM-W](../reference/HDDM_W.html) | Binary streams | Memory efficient |
+| [KSWIN](../reference/KSWIN.html) | Any numeric data | Distribution comparison |
+| [Page-Hinkley](../reference/PageHinkley.html) | Continuous data | Detects mean changes |
+| [KL-Divergence](../reference/KLDivergence.html) | Continuous data | Information-theoretic |
[…]
