### 2026-01-26 37cac47
Adwin added, vignette changed
