+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+export(ADWIN)
 export(DDM)
 export(EDDM)
 export(HDDM_A)
