+++ vignettes/datadriftR-intro.R
@@ -1,0 +1,464 @@
+## ----setup, include = FALSE---------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.width = 7,
+  fig.height = 4,
+  message = FALSE,
+  warning = FALSE
+)
+
+## ----covariate-shift-viz, fig.cap="Covariate Shift: Input distribution changes", fig.height=4----
+set.seed(42)
+par(mfrow = c(1, 2))
+
+# Training data
+x_train <- rnorm(500, mean = 2, sd = 1)
+hist(x_train, breaks = 30, col = "#2A9D8F", border = "white",
+     main = "Training Distribution", xlab = "Feature X", xlim = c(-2, 10))
+
+# Test data (shifted)
+x_test <- rnorm(500, mean = 5, sd = 1.5)
+hist(x_test, breaks = 30, col = "#E63946", border = "white",
+     main = "Production Distribution", xlab = "Feature X", xlim = c(-2, 10))
+
+par(mfrow = c(1, 1))
+
+## ----concept-drift-viz, fig.cap="Concept Drift: Decision boundary shifts", fig.height=4----
+set.seed(123)
+par(mfrow = c(1, 2))
+
+# Before concept drift
+x <- seq(0, 10, length.out = 100)
+y_before <- 0.5 * x + rnorm(100, sd = 0.5)
+plot(x, y_before, pch = 19, col = "#2A9D8F", cex = 0.7,
+     main = "Before Concept Drift", xlab = "X", ylab = "Y")
+abline(lm(y_before ~ x), col = "#1D3557", lwd = 2)
+
+# After concept drift (relationship changed)
+y_after <- -0.3 * x + 8 + rnorm(100, sd = 0.5)
+plot(x, y_after, pch = 19, col = "#E63946", cex = 0.7,
+     main = "After Concept Drift", xlab = "X", ylab = "Y")
+abline(lm(y_after ~ x), col = "#1D3557", lwd = 2)
+
+par(mfrow = c(1, 1))
+
+## ----drift-patterns, fig.cap="Different drift patterns over time", fig.height=5----
+set.seed(456)
+n <- 500
+t <- 1:n
+
+par(mfrow = c(2, 2))
+
+# 1. Abrupt drift
+abrupt <- c(rnorm(250, 0, 1), rnorm(250, 3, 1))
+plot(t, abrupt, type = "l", col = "#457B9D", lwd = 1.5,
+     main = "Abrupt Drift", xlab = "Time", ylab = "Value")
+abline(v = 250, col = "#E63946", lty = 2, lwd = 2)
+
+# 2. Gradual drift
+gradual <- sapply(t, function(i) {
+  p <- min(1, max(0, (i - 200) / 100))
+  rnorm(1, mean = p * 3, sd = 1)
+})
+plot(t, gradual, type = "l", col = "#457B9D", lwd = 1.5,
+     main = "Gradual Drift", xlab = "Time", ylab = "Value")
+rect(200, -5, 300, 8, col = rgb(0.9, 0.3, 0.3, 0.2), border = NA)
+
+# 3. Incremental drift
+incremental <- rnorm(n, mean = t/100, sd = 1)
+plot(t, incremental, type = "l", col = "#457B9D", lwd = 1.5,
+     main = "Incremental Drift", xlab = "Time", ylab = "Value")
+
+# 4. Recurring drift (seasonal)
+recurring <- sin(t / 30) * 2 + rnorm(n, 0, 0.5)
+plot(t, recurring, type = "l", col = "#457B9D", lwd = 1.5,
+     main = "Recurring/Seasonal Drift", xlab = "Time", ylab = "Value")
+
[…]
