+++ vignettes/datadriftR-intro.Rmd
@@ -1,8 +1,8 @@
 ---
-title: "Comprehensive Guide to Drift Detection with datadriftR"
+title: "Getting Started with datadriftR"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Comprehensive Guide to Drift Detection with datadriftR}
+  %\VignetteIndexEntry{Getting Started with datadriftR}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -20,696 +20,319 @@ knitr::opts_chunk$set(
 
 # Introduction
 
-Machine learning models are trained on historical data with the assumption that future data will follow similar patterns. However, in real-world applications, **data evolves over time**. When the statistical properties of data change, model performance can silently degrade—leading to incorrect predictions, financial losses, or even safety risks.
+Machine learning models assume that future data follows similar patterns to training data. In production, **data evolves over time**—and when statistical properties change, model performance silently degrades.
 
-**datadriftR** provides a comprehensive toolkit for detecting these changes in streaming data, enabling proactive model maintenance and reliable production ML systems.
+**datadriftR** detects these changes in streaming data, enabling proactive model maintenance.
 
-# Understanding Drift: A Conceptual Framework
+# Types of Drift
 
-## What is Dataset Shift?
-
-**Dataset shift** is the umbrella term for any scenario where the joint distribution of inputs and outputs differs between training and deployment:
-
-$$P_{train}(X, Y) \neq P_{test}(X, Y)$$
-
-This shift can manifest in several distinct forms, each requiring different detection strategies.
-
-## Types of Drift
-
-### 1. Covariate Shift (Data Drift / Virtual Drift)
-
-The input distribution changes while the relationship between inputs and outputs remains the same:
-
-$$P_{train}(X) \neq P_{test}(X) \quad \text{but} \quad P(Y|X) \text{ unchanged}$$
-
-**Real-world examples:**
-
-- **E-commerce**: Customer demographics shift from urban millennials to suburban families
-- **Healthcare**: Hospital starts serving older patient population
-- **Finance**: Economic recession changes spending patterns
-
-```{r covariate-shift-viz, fig.cap="Covariate Shift: Input distribution changes", fig.height=4}
+```{r drift-types-viz, fig.height=5, fig.cap="Common drift patterns"}
 set.seed(42)
-par(mfrow = c(1, 2))
+n <- 400
+par(mfrow = c(2, 2), mar = c(4, 4, 3, 1))
 
-# Training data
-x_train <- rnorm(500, mean = 2, sd = 1)
-hist(x_train, breaks = 30, col = "#2A9D8F", border = "white",
-     main = "Training Distribution", xlab = "Feature X", xlim = c(-2, 10))
-
-# Test data (shifted)
-x_test <- rnorm(500, mean = 5, sd = 1.5)
-hist(x_test, breaks = 30, col = "#E63946", border = "white",
-     main = "Production Distribution", xlab = "Feature X", xlim = c(-2, 10))
-
-par(mfrow = c(1, 1))
-```
-
-### 2. Concept Drift (Real Drift)
-
-The relationship between inputs and outputs changes:
-
[…]
