+++ paper/paper.md
@@ -28,13 +28,13 @@ affiliations:
 
 # Summary
 
-Deployed machine learning models frequently encounter degradation in predictive accuracy when the statistical properties of incoming data evolve over time, a condition known as data drift. This phenomenon can manifest in several forms, most notably concept drift, which occurs when the functional relationship linking predictor variables to the outcome changes, thereby undermining model reliability. Conventional drift detection strategies often rely on aggregate performance indicators or univariate distributional summaries, approaches that may overlook nuanced yet consequential shifts in the data-generating mechanism. Within the broader machine learning operations (MLOps) framework, continuous model monitoring has emerged as a critical practice for safeguarding the stability and dependability of production systems [@biecek2019; @mougan2023]. datadriftR is an open-source R package designed to address these challenges by providing real-time detection of data drift in univariate streaming data. The package implements a comprehensive suite of widely recognized statistical methods for monitoring distributional changes, including error-rate-based detectors that track classification performance (DDM [@gama2004], EDDM [@baena2006]), Hoeffding-bound methods that employ adaptive windowing to detect mean shifts (HDDM-A and HDDM-W [@frias2014]), a sliding-window Kolmogorov–Smirnov test for distribution comparison (KSWIN [@raab2020]), the cumulative-sum-based Page–Hinkley test for detecting persistent shifts [@page1954], histogram-based Kullback–Leibler divergence monitoring for measuring distributional divergence [@kullback1951], and a functional profile comparison method for analyzing temporal patterns [@kobylinska2023]. 
+Deployed machine learning models frequently encounter degradation in predictive accuracy when the statistical properties of incoming data evolve over time, a condition known as data drift. This phenomenon can manifest in several forms, most notably concept drift, which occurs when the functional relationship linking predictor variables to the outcome changes, thereby undermining model reliability. Conventional drift detection strategies often rely on aggregate performance indicators or univariate distributional summaries, approaches that may overlook nuanced yet consequential shifts in the data-generating mechanism. Within the broader machine learning operations (MLOps) framework, continuous model monitoring has emerged as a critical practice for safeguarding the stability and dependability of production systems [@biecek2019; @mougan2023]. datadriftR is an open-source R package designed to address these challenges by providing real-time detection of data drift in univariate streaming data. The package implements a comprehensive suite of widely recognized statistical methods for monitoring distributional changes, including error-rate-based detectors that track classification performance (DDM [@gama2004], EDDM [@baena2006]), Hoeffding-bound methods that employ adaptive windowing to detect mean shifts (HDDM-A and HDDM-W [@frias2014]), adaptive windowing for change detection (ADWIN [@bifet2007adwin]), a sliding-window Kolmogorov–Smirnov test for distribution comparison (KSWIN [@raab2020]), the cumulative-sum-based Page–Hinkley test for detecting persistent shifts [@page1954], histogram-based Kullback–Leibler divergence monitoring for measuring distributional divergence [@kullback1951], and a functional profile comparison method for analyzing temporal patterns [@kobylinska2023]. 
 
 # Statement of need
 
 Data drift detection is a fundamental challenge in deployed machine learning systems and adaptive analytics [@kobylinska2023]. When the underlying data-generating process changes over time, model performance can deteriorate silently, leading to incorrect predictions and suboptimal decision-making. Early detection of such shifts enables timely interventions—such as model retraining, recalibration or triggering alerts—thereby maintaining system reliability in production environments.
 
[…]
