+++ paper/paper.bib
@@ -115,3 +115,13 @@
   year         = {2008},
   publisher    = {Springer}
 }
+
+@inproceedings{bifet2007adwin,
+  title        = {Learning from Time-Changing Data with Adaptive Windowing},
+  author       = {Bifet, Albert and Gavald{\'a}, Ricard},
+  booktitle    = {Proceedings of the 2007 SIAM International Conference on Data Mining},
+  pages        = {443--448},
+  year         = {2007},
+  publisher    = {Society for Industrial and Applied Mathematics},
+  doi          = {10.1137/1.9781611972771}
+}
