### 2026-01-27 8bd176e
small changes in paper, citation and description
