+++ CITATION.cff
@@ -6,7 +6,7 @@ version: "1.1.0"
 date-released: "2025-11-09"
 repository-code: "https://github.com/ugurdar/datadriftR"
 url: "https://cran.r-project.org/package=datadriftR"
-license: "GPL-2.0-or-later"
+license: "MIT"
 keywords:
   - R
   - concept drift
@@ -23,10 +23,11 @@ authors:
     email: mustafacavus@eskisehir.edu.tr
     orcid: "https://orcid.org/0000-0002-6172-5449"
 abstract: |
-  datadriftR is an open-source R package that implements a suite of 
-  concept drift detection algorithms for univariate data streams, including
-  DDM, EDDM, HDDM-A, HDDM-W, KSWIN, Page–Hinkley, ADWIN and KL divergence–based
-    monitoring. The package leverages efficient R6 classes for real-time monitoring and adaptive online learning scenarios.
+  datadriftR provides concept drift and change detection for streaming data in
+  R, implementing established methods including DDM, EDDM, HDDM-A, HDDM-W,
+  KSWIN, Page–Hinkley, ADWIN and KL divergence–based monitoring. It supports
+  online learning workflows and model monitoring by detecting shifts in data
+  distributions and predictive performance over time.
 
 preferred-citation:
   type: software
