+++ DESCRIPTION
@@ -41,7 +41,8 @@ Suggests:
     pkgdown,
     dynaTree,
     ranger
+Config/testthat/edition: 3
 VignetteBuilder: knitr
 RoxygenNote: 7.3.2
-URL: https://github.com/ugurdar/datadriftR
+URL: https://ugurdar.github.io/datadriftR/, https://github.com/ugurdar/datadriftR
 BugReports: https://github.com/ugurdar/datadriftR/issues
