+++ articles/datadriftR-intro.md
@@ -1,0 +1,375 @@
+# Drift Detection with datadriftR
+
+## Introduction
+
+In production, the data generating process rarely stays still.
+Seasonality, product changes, user behavior, and policy updates can
+shift the distribution of your features—and sometimes even change the
+relationship between features and the target.
+
+**Example (concept drift):** You train a model to predict whether
+electricity prices will go up based on demand and price signals. After a
+market reform, the same demand level can imply a different price
+movement. In practice this often shows up as a rising error rate or
+shifting model outputs.
+
+**datadriftR** helps you detect these shifts early in streaming
+settings, so you can trigger investigation, retraining, or alerts.
+
+## Drift Taxonomy
+
+Many “drift” problems are forms of **dataset shift**:
+
+$$P_{train}(X,Y) \neq P_{prod}(X,Y)$$
+
+What changed?
+
+- **Covariate shift:** $P(X)$ changes while $P\left( Y|X \right)$ stays
+  roughly the same.
+- **Concept drift:** $P\left( Y|X \right)$ changes (the mapping from
+  inputs to outputs changes).
+
+## Drift Patterns
+
+Drift can also differ by **how it unfolds over time**:
+
+- **Abrupt:** change happens at once (deployments, outages, policy
+  changes)
+- **Gradual:** mixture of old/new concepts (slow adoption, rollout)
+- **Incremental:** continuously moving distribution (wear-and-tear,
+  trend)
+- **Recurring:** seasonal or cyclical patterns (day/night, weekdays,
+  holidays)
+
+![Common drift
+patterns](datadriftR-intro_files/figure-html/drift-types-viz-1.png)
+
+Common drift patterns
+
+## Available Methods
+
+| Method                                        | What you feed                              |
+|:----------------------------------------------|:-------------------------------------------|
+| **DDM, EDDM, HDDM-A/W**                       | Binary (0/1) error stream                  |
+| **KSWIN, ADWIN, Page-Hinkley, KL Divergence** | Numeric stream (or rolling window/batches) |
+| **ProfileDifference**                         | Two profiles (e.g., PDPs)                  |
+
+------------------------------------------------------------------------
+
+## Quick Start
+
+Quick start with
+[`detect_drift()`](https://ugurdar.github.io/datadriftR/reference/detect_drift.md):
+
+``` r
+set.seed(1)
+x <- c(rnorm(300, 0, 1), rnorm(200, 3, 1))
+
+detect_drift(x, method = "page_hinkley", delta = 0.05, threshold = 50)
+#>   index    value  type
+#> 1   315 2.964078 drift
+```
+
+------------------------------------------------------------------------
+
+## Part 1: Error-Rate Detectors (Binary Stream)
+
+These methods monitor a stream of **prediction errors**:
+
+``` r
+error_stream <- as.integer(y_pred != y_true)
+```
+
+They work well when you can observe labels reasonably quickly (or with a
+known delay).
+
+### Simulating Model Errors
+
+``` r
+set.seed(123)
+
+n_good <- 500
+n_bad <- 500
+error_stream <- c(
+  rbinom(n_good, 1, prob = 0.05),
+  rbinom(n_bad, 1, prob = 0.30)
+)
+true_drift_error <- n_good + 1
+```
+
+### Comparing All Error-Rate Methods
+
+``` r
+error_methods <- c("ddm", "eddm", "hddm_a", "hddm_w")
+
+first_index <- function(res, type) {
+  idx <- res$index[res$type == type]
[…]
