+++ reference/HDDM_A.md
@@ -1,0 +1,279 @@
+# HDDM_A: Drift Detection Method based on Adaptive Windows
+
+This class implements the HDDM_A drift detection method that uses
+adaptive windows to detect changes in the mean of a data stream. It is
+designed to monitor online streams of data and can detect increases or
+decreases in the process mean in a non-parametric and online manner.
+
+## Details
+
+HDDM_A adapts to changes in the data stream by adjusting its internal
+windows to track the minimum and maximum values of the process mean. It
+triggers alerts when a significant drift from these benchmarks is
+detected.
+
+## References
+
+Frías-Blanco I, del Campo-Ávila J, Ramos-Jimenez G, et al. Online and
+non-parametric drift detection methods based on Hoeffding’s bounds. IEEE
+Transactions on Knowledge and Data Engineering, 2014, 27(3): 810-823.
+
+Albert Bifet, Geoff Holmes, Richard Kirkby, Bernhard Pfahringer. MOA:
+Massive Online Analysis; Journal of Machine Learning Research 11:
+1601-1604, 2010.
+
+Implementation:
+https://github.com/scikit-multiflow/scikit-multiflow/blob/a7e316d1cc79988a6df40da35312e00f6c4eabb2/src/skmultiflow/drift_detection/hddm_a.py
+
+## Public fields
+
+- `drift_confidence`:
+
+  Confidence level for detecting a drift.
+
+- `warning_confidence`:
+
+  Confidence level for warning detection.
+
+- `two_side_option`:
+
+  Boolean flag for one-sided or two-sided mean monitoring.
+
+- `total_n`:
+
+  Total number of samples seen.
+
+- `total_c`:
+
+  Total cumulative sum of the samples.
+
+- `n_max`:
+
+  Maximum window end for sample count.
+
+- `c_max`:
+
+  Maximum window end for cumulative sum.
+
+- `n_min`:
+
+  Minimum window start for sample count.
+
+- `c_min`:
+
+  Minimum window start for cumulative sum.
+
+- `n_estimation`:
+
+  Number of samples since the last detected change.
+
+- `c_estimation`:
+
+  Cumulative sum since the last detected change.
+
+- `change_detected`:
+
+  Boolean indicating if a change was detected.
+
+- `warning_detected`:
+
+  Boolean indicating if a warning has been detected.
+
+- `estimation`:
+
+  Current estimated mean of the stream.
+
+- `delay`:
+
+  Current delay since the last update.
+
+## Methods
+
+### Public methods
+
+- [`HDDM_A$new()`](#method-HDDM_A-new)
+
+- [`HDDM_A$add_element()`](#method-HDDM_A-add_element)
+
+- [`HDDM_A$mean_incr()`](#method-HDDM_A-mean_incr)
+
+- [`HDDM_A$mean_decr()`](#method-HDDM_A-mean_decr)
+
+- [`HDDM_A$reset()`](#method-HDDM_A-reset)
+
+- [`HDDM_A$update_estimations()`](#method-HDDM_A-update_estimations)
+
+- [`HDDM_A$clone()`](#method-HDDM_A-clone)
+
+------------------------------------------------------------------------
+
+### Method `new()`
+
+Initializes the HDDM_A detector with specific settings.
+
+#### Usage
+
+    HDDM_A$new(
+      drift_confidence = 0.001,
+      warning_confidence = 0.005,
+      two_side_option = TRUE
+    )
+
+#### Arguments
+
+- `drift_confidence`:
+
+  Confidence level for drift detection.
+
+- `warning_confidence`:
+
+  Confidence level for issuing warnings.
+
+- `two_side_option`:
+
+  Whether to monitor both increases and decreases.
+
+------------------------------------------------------------------------
+
[…]
