+++ articles/index.md
@@ -1,0 +1,6 @@
+# Articles
+
+### All vignettes
+
+- [Drift Detection with
+  datadriftR](https://ugurdar.github.io/datadriftR/articles/datadriftR-intro.md):
