### 2026-01-27 d03b24d
KSWIN example added
