+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.2.0
 pkgdown_sha: ~
 articles:
   datadriftR-intro: datadriftR-intro.html
-last_built: 2026-01-27T17:57Z
+last_built: 2026-01-27T18:13Z
 urls:
   reference: https://ugurdar.github.io/datadriftR/reference
   article: https://ugurdar.github.io/datadriftR/articles
