### 2026-01-27 ugurdar/datadriftR#9 @ugurdar
@fangzhou-xie  Thank you for the review and for highlighting this important points. We appreciate the opportunity to clarify the scope of our contribution and how the package has evolved in response to reviewers feedback.

## 1. Substantial scholarly effort

We agree that several detectors in datadriftR implement established algorithms that also exist in other ecosystems (e.g., `scikit-multiflow`); we do not claim algorithmic novelty for those methods. Our contribution is to make these widely used drift detectors available in R as a coherent, documented, and tested package with a consistent interface.

Importantly, based on reviewers feedback about R‑native usability, we added the high-level wrapper `detect_drift()`, which processes a whole stream vector and returns detections as a `data.frame` (rather than requiring user-written loops and console messages). In addition, the package includes distribution- and monitoring components such as `KL-divergence`, `ADWIN` (recently added), and `ProfileDifference`, which broaden the scope beyond a direct translation of a single module. We defer to the editor regarding the final interpretation of the “substantial scholarly effort” criterion.

## 2. Installation issues

We updated the GitHub install path to the correct repository (ugurdar/datadriftR) and ensured the README includes the standard CRAN installation via `install.packages("datadriftR")`.

Missing dependencies in Imports: We confirmed the issue and fixed it by adding the required packages (`fda.usc` and `doremi`) to the Imports field in [DESCRIPTION](https://github.com/ugurdar/datadriftR/blob/0b537cf427464943a2092622d068c0fa954a1ebf/DESCRIPTION#L35), so installation from GitHub no longer fails due to undeclared dependencies.

We synchronized the development version in the repository and set it to 1.1.0. The current CRAN release is 1.0.0, and we plan to submit an updated CRAN release to align CRAN with the repository and the latest fixes.

## 3. Functionality

Thank you for this perspective—we agree with the core point. While the underlying detectors are naturally stateful and can be updated observation-by-observation (which is useful for truly streaming integrations), many R users work with vectors/data frames and expect functions that return results directly.

In response, we provide an R-friendly, high-level interface ([`detect_drift()`](https://github.com/ugurdar/datadriftR/blob/paper/R/detect_drift.R)) that takes a stream as a vector and returns detections as a structured output (data.frame with indices/values and drift vs. warning when applicable), so users do not need to write explicit `for` loops. At the same time, we keep the original “online” usage via the R6 detector classes (e.g., `DDM$new()`, `PageHinkley$new()`) as an alternative mode for applications that require incremental updates and persistent detector state.

## 4. Documentation

We updated the installation instructions and corrected the `DESCRIPTION` so all required dependencies are properly declared under [Imports](https://github.com/ugurdar/datadriftR/blob/0b537cf427464943a2092622d068c0fa954a1ebf/DESCRIPTION#L32), resolving the GitHub installation failure.
We set up and published a [pkgdown](https://ugurdar.github.io/datadriftR) site; the documentation is available (vignettes + reference pages).
We added a complete usage example for KSWIN .

We added a test suite using [testthat](https://github.com/ugurdar/datadriftR/tree/paper/tests) and covered all implemented detectors. In particular, we added/expanded tests for [`DDM`](https://github.com/ugurdar/datadriftR/blob/0b537cf427464943a2092622d068c0fa954a1ebf/tests/testthat/test-ddm.R#L69), [`EDDM`](https://github.com/ugurdar/datadriftR/blob/0b537cf427464943a2092622d068c0fa954a1ebf/tests/testthat/test-eddm.R#L49), [`HDDM-A`](https://github.com/ugurdar/datadriftR/blob/0b537cf427464943a2092622d068c0fa954a1ebf/tests/testthat/test-hddm.R#L77), and [`HDDM-W`](https://github.com/ugurdar/datadriftR/blob/0b537cf427464943a2092622d068c0fa954a1ebf/tests/testthat/test-hddm.R#L117), and verified that the detected drift indices match  the reference behavior reported by scikit-multiflow on the same test streams. We did not include the Python reference scripts to avoid introducing Python into the R package workflow, but we can provide them under a separate folder (e.g., tools/python-reference-tests/) if preferred.

## 5. Misc

We agree that wording was misleading. We corrected the dependency declarations in `DESCRIPTION` (including `fda.usc` and `doremi` ) and removed the “lightweight / base R + R6 only” style claim from the documentation.

Return format / usability: We agree that structured outputs are preferable in R. We added `detect_drift()`, which returns detections as a data.frame (including indices/values and drift vs. warning when applicable), so users do not need to rely on console messages or write explicit loops.
