### 2026-01-27 ugurdar/datadriftR#9 @fangzhou-xie
Thank you for your work on this package. I have checked and can confirm that all my comments were addressed.

One more tip: you could also consider adding the documentation URL to the "About" section of Github repo. It is currently saying "No description, website, or topics provided." Putting the documentation webpage there would make people's lives easier to access the docs.

Given that my concerns were all addressed, I will finish my report at the JOSS repo soon.
