+++ .github/workflows/R-CMD-check.yaml
@@ -1,4 +1,3 @@
-  - uses: r-lib/actions/setup-pandoc@v2
 name: R-CMD-check
 
 on:
@@ -11,12 +10,15 @@ jobs:
   R-CMD-check:
     runs-on: ${{ matrix.os }}
     strategy:
+      fail-fast: false
       matrix:
         os: [ubuntu-latest, windows-latest, macos-latest]
         r-version: ['4.1', 'release']
     
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
+
+      - uses: r-lib/actions/setup-pandoc@v2
       
       - uses: r-lib/actions/setup-r@v2
         with:
@@ -27,10 +29,6 @@ jobs:
         run: |
           sudo apt-get update
           sudo apt-get install -y libcurl4-openssl-dev libssl-dev libxml2-dev
-      
-      - name: Install Pandoc (macOS)
-        if: runner.os == 'macOS'
-        run: brew install pandoc
 
       - name: Install R dependencies
         run: |
