+++ datadriftR.Rcheck/00_pkg_src/datadriftR/paper/paper.bib
@@ -1,0 +1,127 @@
+@inproceedings{gama2004,
+  title        = {Learning with Drift Detection},
+  author       = {Gama, Jo\~{a}o and Medas, Pedro and Castillo, Gladys and Rodrigues, Pedro},
+  booktitle    = {Advances in Artificial Intelligence -- SBIA 2004},
+  series       = {Lecture Notes in Computer Science},
+  year         = {2004}
+}
+
+@inproceedings{baena2006,
+  title        = {Early Drift Detection Method},
+  author       = {Baena-Garc{\'\i}a, Manuel and del Campo-{\'A}vila, Jos{\'e} and Fidalgo, Joaqu{\'\i}n and Bifet, Albert and Gavald{\'a}, Ricard and Morales-Bueno, Rafael},
+  booktitle    = {ECML PKDD 2006 Workshop on Knowledge Discovery from Data Streams},
+  year         = {2006}
+}
+
+@article{frias2014,
+  title        = {Online and Nonparametric Drift Detection Methods Based on Hoeffding's Bounds},
+  author       = {Fr{\'{\i}}as-Blanco, Isvani and del Campo-{\'A}vila, Jos{\'e} and Ramos-Jim{\'e}nez, Gustavo and Morales-Bueno, Rafael and Ortiz-D{\'{\i}}az, Eric and Caballero-Mota, Yudith},
+  journal      = {IEEE Transactions on Knowledge and Data Engineering},
+  volume       = {27},
+  number       = {3},
+  pages        = {810--823},
+  year         = {2015},
+  doi          = {10.1109/TKDE.2014.2345382}
+}
+
+@article{raab2020,
+  title        = {REDC: Regularized Drift Detection for MOA},
+  author       = {Raab, M. and Heusinger, M. and Schleif, F.-M.},
+  journal      = {Neurocomputing},
+  year         = {2020},
+  doi          = {10.1016/j.neucom.2019.11.111}
+}
+
+@article{kullback1951,
+  title        = {On Information and Sufficiency},
+  author       = {Kullback, Solomon and Leibler, Richard A.},
+  journal      = {The Annals of Mathematical Statistics},
+  volume       = {22},
+  number       = {1},
+  pages        = {79--86},
+  year         = {1951},
+  doi          = {10.1214/aoms/1177729694}
+}
+
+@article{page1954,
+  title        = {Continuous Inspection Schemes},
+  author       = {Page, E. S.},
+  journal      = {Biometrika},
+  volume       = {41},
+  number       = {1-2},
+  pages        = {100--115},
+  year         = {1954},
+  doi          = {10.1093/biomet/41.1-2.100}
+}
+
+@article{montiel2018,
+  title        = {Scikit-Multiflow: A Multi-output Streaming Framework},
+  author       = {Montiel, Jacob and Read, Jesse and Bifet, Albert and Abdessalem, Talel},
+  journal      = {Journal of Machine Learning Research},
+  volume       = {19},
+  number       = {72},
+  pages        = {1--5},
+  year         = {2018},
+  url          = {https://jmlr.org/papers/v19/18-251.html}
+}
+
+@misc{kobylinska2023,
+  title        = {A Survey on Concept Drift Adaptation},
+  author       = {Kobyli{\'n}ska, Ewa and others},
+  year         = {2023},
+  eprint       = {2308.11446},
+  archivePrefix= {arXiv},
+  primaryClass = {cs.LG},
+  url          = {https://arxiv.org/abs/2308.11446}
+}
+
+@article{bifet2010moa,
+  title        = {MOA: Massive Online Analysis},
+  author       = {Bifet, Albert and Holmes, Geoff and Kirkby, Richard and Pfahringer, Bernhard},
[…]
