+++ datadriftR.Rcheck/tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(datadriftR)
+
+test_check("datadriftR")
