+++ datadriftR.Rcheck/00_pkg_src/datadriftR/R/ProfileDifference.R
@@ -1,0 +1,194 @@
+#' Profile Difference Calculation for Change Detection
+#'
+#' @description
+#' Implements the calculation of profile differences using various methods such as PDI, L2, and L2 derivative.
+#' The class provides methods for setting profiles and calculating the differences.
+#'
+#' @details
+#' The class supports multiple methods for calculating profile differences, including the Profile Disparity Index (PDI)
+#' using gold or simple derivative methods, and L2 norm and L2 derivative calculations. It allows for customization
+#' of various parameters such as embedding dimensions, derivative orders, and thresholds.
+#'
+#' @references
+#' Kobylińska, K., Krzyziński, M., Machowicz, R., Adamek, M., & Biecek, P. (2023). Exploration of the Rashomon Set Assists Trustworthy Explanations for Medical Data. arXiv e-prints, arXiv-2308.
+#'
+#' @examples
+#' set.seed(123)
+#' profile1 <- list(x = 1:100, y = sin(1:100))
+#' profile2 <- list(x = 1:100, y = sin(1:100) + rnorm(100, 0, 0.1))
+#' pd <- ProfileDifference$new(method = "pdi", deriv = "gold")
+#' pd$set_profiles(profile1, profile2)
+#' result <- pd$calculate_difference()
+#' message("Method:", result$method)
+#' message("Distance:", round(result$distance, 4))
+#' @import R6
+#' @import fda.usc
+#' @import doremi
+#' @import stats
+#' @export
+ProfileDifference <- R6Class(
+  "ProfileDifference",
+  public = list(
+    #' @field method The method used for profile difference calculation.
+    method = "pdi",
+    #' @field deriv The method used for derivative calculation.
+    deriv = "gold",
+    #' @field gold_spline Boolean indicating if cubic spline should be used in gold method.
+    gold_spline = TRUE,
+    #' @field gold_embedding Embedding dimension for gold method.
+    gold_embedding = 4,
+    #' @field nderiv Order of the derivative for simple method.
+    nderiv = 4,
+    #' @field gold_spline_threshold Threshold for cubic spline in gold method.
+    gold_spline_threshold = 0.01,
+    #' @field epsilon Small value to avoid numerical issues.
+    epsilon = NULL,
+    #' @field profile1 The first profile.
+    profile1 = NULL,
+    #' @field profile2 The second profile.
+    profile2 = NULL,
+
+    #' @description
+    #' Initializes the ProfileDifference class.
+    #' @param method The method used for profile difference calculation.
+    #' @param deriv The method used for derivative calculation.
+    #' @param gold_spline Boolean indicating if cubic spline should be used in gold method.
+    #' @param gold_embedding Embedding dimension for gold method.
+    #' @param nderiv Order of the derivative for simple method.
+    #' @param gold_spline_threshold Threshold for cubic spline in gold method.
+    #' @param epsilon Small value to avoid numerical issues.
+    initialize = function(method = "pdi",
+                          deriv = "gold",
+                          gold_spline = TRUE,
+                          gold_embedding = 4,
+                          nderiv = 4,
+                          gold_spline_threshold = 0.01,
+                          epsilon = NULL) {
+      self$method <- method
+      self$deriv <- deriv
+      self$gold_spline <- gold_spline
+      self$gold_embedding <- gold_embedding
+      self$nderiv <- nderiv
+      self$gold_spline_threshold <- gold_spline_threshold
+      self$epsilon <- epsilon
+      self$reset()
+    },
+
+    #' @description
+    #' Resets the internal state of the detector.
+    reset = function() {
+      self$profile1 <- NULL
+      self$profile2 <- NULL
+    },
+
[…]

[…]
