+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.2.0
 pkgdown_sha: ~
 articles:
   datadriftR-intro: datadriftR-intro.html
-last_built: 2026-01-30T22:21Z
+last_built: 2026-01-30T23:05Z
 urls:
   reference: https://ugurdar.github.io/datadriftR/reference
   article: https://ugurdar.github.io/datadriftR/articles
