### 2026-02-03 #9481 @benyamindsmith
Thank you @ugurdar for addressing my issues.

I have no further concerns and fully endorse `datadriftR`  as a useful package for data drift detection for both educational and production uses.  I hope to get a chance to work with it as a consumer in my own research/work.

@crvernon I recommend that JOSS accepts and publishes the associated paper.

Thank you for giving me the opportunity to review the work.
