### 2026-02-03 #9481 @ugurdar
Hi @crvernon — should I do anything on my side right now?
Also, I changed the paper title from “datadriftR: an R package for streaming data drift” to [datadriftR: An R package for data drift detection"](https://github.com/ugurdar/datadriftR/blob/8099168808c72eafe4eee8fa50afb741cf705315/paper/paper.md?plain=1#L2) in our paper.
