### 2026-03-27 #9481 @ugurdar
> 👋 [@ugurdar](https://github.com/ugurdar) - Your paper is in great shape so we are almost there! Next is just setting up the archive for your new release.
>
> We want to make sure the archival has the correct metadata that JOSS requires. This includes a title that matches the paper title and a correct author list.
>
> So here is what we have left to do:
>
>     * [ ]
>       [x]   Confirm that the title of your paper in your paper.md file has no special formatting (e.g., backticks) as this will cause formatting issues when accepted.
>       [ ]
>       [x]   Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository).  Please ensure that the software archive uses the same license as the license you have posted in your code repository.
>       [ ]
>       [x]   Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly including the case) and the author list (should match the people and the order in which they are listed on your paper exactly).
>       [ ]
>       [x]   Please respond with the DOI of the archived version here along with the version tag name released.
>
>
> I can then move forward with accepting the submission.

Hi @crvernon thank you for the feedback. This is DOI of https://doi.org/10.5281/zenodo.19262539 the Zenodo repo.
