### 2026-03-30 #9481 @crvernon
🥳 Congratulations on your new publication @ugurdar ! Many thanks to @fangzhou-xie and @benyamindsmith for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
